Changed the command name to inspect, changed cache-hit command to inspect db:cache-hit
