+++ internal/inspect/cache/cache.go
